+++ src/Rocketeer/Tasks/Test.php
@@ -1,0 +1,24 @@
+<?php
+namespace Rocketeer\Tasks;
+
+use Rocketeer\Tasks\Abstracts\Task;
+
+/**
+ * Run the tests on the server and displays the ouput
+ */
+class Test extends Task
+{
+
+	/**
+	 * Run the Task
+	 *
+	 * @return  void
+	 */
+	public function execute()
+	{
+		// Update repository
+		$this->command->info('Testing the application');
+		$this->runTests();
+	}
+
+}