+++ tests/_start.php
@@ -36,6 +36,8 @@ abstract class RocketeerTests extends PHPUnit_Framework_TestCase
 
 	/**
 	 * Set up the tests
+	 *
+	 * @return void
 	 */
 	public function setUp()
 	{
@@ -49,6 +51,11 @@ abstract class RocketeerTests extends PHPUnit_Framework_TestCase
 		$command = $this->getCommand();
 		$config  = $this->getConfig();
 		$remote  = $this->getRemote();
+
+		// Laravel classes --------------------------------------------- /
+
+		$this->app['path.base']    = '/src';
+		$this->app['path.storage'] = '/src/storage';
 
 		$this->app->singleton('config', function() use ($config) {
 			return $config;
@@ -64,7 +71,8 @@ abstract class RocketeerTests extends PHPUnit_Framework_TestCase
 
 		// Rocketeer classes ------------------------------------------- /
 
-		$this->app = RocketeerServiceProvider::bindClasses($this->app);
+		$serviceProvider = new RocketeerServiceProvider($this->app);
+		$this->app = $serviceProvider->bindClasses($this->app);
 
 		$this->app->bind('rocketeer.deployments', function($app) {
 			return new Rocketeer\DeploymentsManager($app['files'], __DIR__);
@@ -78,6 +86,11 @@ abstract class RocketeerTests extends PHPUnit_Framework_TestCase
 		$this->task = $this->task('Cleanup');
 	}
 
+	/**
+	 * Recreate placeholder server
+	 *
+	 * @return void
+	 */
 	public function tearDown()
 	{
 		// Recreate deployments file
@@ -142,6 +155,8 @@ abstract class RocketeerTests extends PHPUnit_Framework_TestCase
 		$command->shouldReceive('line')->andReturnUsing(function($message) { return $message; });
 		$command->shouldReceive('info')->andReturnUsing(function($message) { return $message; });
 		$command->shouldReceive('argument');
+		$command->shouldReceive('ask');
+		$command->shouldReceive('secret');
 		if ($option) $command->shouldReceive('option');
 
 		return $command;
@@ -155,11 +170,15 @@ abstract class RocketeerTests extends PHPUnit_Framework_TestCase
 	protected function getConfig()
 	{
 		$config = Mockery::mock('Illuminate\Config\Repository');
+		$config->shouldReceive('get')->with('database.default')->andReturn('mysql');
+		$config->shouldReceive('get')->with('cache.driver')->andReturn('file');
+		$config->shouldReceive('get')->with('session.driver')->andReturn('file');
 		$config->shouldReceive('get')->with('rocketeer::remote.application_name')->andReturn('foobar');
 		$config->shouldReceive('get')->with('rocketeer::remote.root_directory')->andReturn(__DIR__.'/server/');
 		$config->shouldReceive('get')->with('rocketeer::remote.keep_releases')->andReturn(1);
 		$config->shouldReceive('get')->with('rocketeer::remote.shared')->andReturn(array('tests/meta'));
 		$config->shouldReceive('get')->with('rocketeer::git.branch')->andReturn('master');
+		$config->shouldReceive('get')->with('rocketeer::git.repository')->andReturn('https://github.com/Anahkiasen/rocketeer.git');
 		$config->shouldReceive('get')->with('rocketeer::connections')->andReturn('production');
 
 		$config->shouldReceive('get')->with('rocketeer::tasks')->andReturn(array(