+++ CHANGELOG.md
@@ -3,8 +3,20 @@
 ### 0.x.x (roadmap)
 
 - Ability to select which severs a Task executes on, on a per-task basis
+- Add compatibility to Laravel 4.0
 
-### 0.4.0 (develop)
+### 0.5.0 (stable)
+
+- **Add a `deploy:update` task that updates the remote server without doing a new release**
+- **Add a `deploy:test` to run the tests on the server**
+- **Rocketeer can now prompt for Git credentials if you don't want to store them in the config**
+- The `deploy:check` command now checks PHP extensions for the cache/database/session drivers you set
+- Rocketeer now share logs by default between releases
+- Add ability to specify an array of Tasks in Rocketeer::before|after
+- Added a `$silent` flag to make a `Task::run` call silent no matter what
+- Rocketeer now displays how long the task took
+
+### 0.4.0
 
 - **Add ability to share files and folders between releases**
 - **Add ability to create custom tasks integrated in the CLI**
@@ -14,7 +26,7 @@
 - Add `Rocketeer::add` method on the facade to register custom Tasks
 - Fix `Task::runComposer` not taking into account a local `composer.phar`
 
-### 0.3.2 (stable)
+### 0.3.2
 
 - Fixed wrong tag used in `deploy:cleanup`